+++ src/test/java/bowling/FrameTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.Frame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
